chore(tests): fold integration-upgrade tests into test_integration and simplify CI
