+++ .github/workflows/ci.yml
@@ -44,7 +44,6 @@ jobs:
     strategy:
       matrix:
         python: ["3.10", "3.11", "3.12", "3.13"]
-        test_target: ["full", "integration-upgrade"]
 
     steps:
     - name: Checkout repository
@@ -70,12 +69,7 @@ jobs:
 
     - name: Run tests
       shell: bash
-      run: |
-        if [[ "${{ matrix.test_target }}" == "integration-upgrade" ]]; then
-          pytest tests/test_integration_upgrade.py
-        else
-          pytest
-        fi
+      run: pytest
 
   # Tag and release the package
   release:
